Support basic values
Adds support to use basic values for some fields such as category
to be able to provide an instance of NSNumber for option ID
instead of a full PKTCategoryOption instance. For date fields, it
now supports only providing an NSDate instance which will be used
as the start date of the field, instead of a PKTDateRange value.